dashboard/requirements.txt
streamlit==1.56.0
google-cloud-bigquery==3.41.0
db-dtypes==1.5.1
plotly==6.7.0
pandas==2.3.3
PyYAML==6.0.3
python-dateutil==2.9.0.post0